+++ dds_web/api/__init__.py
@@ -61,6 +61,7 @@ api.add_resource(project.ProjectUsers, "/proj/users", endpoint="list_project_use
 api.add_resource(project.ProjectStatus, "/proj/status", endpoint="project_status")
 api.add_resource(project.ProjectAccess, "/proj/access", endpoint="project_access")
 api.add_resource(project.ProjectBusy, "/proj/busy", endpoint="project_busy")
+api.add_resource(project.ProjectInfo, "/proj/info", endpoint="project_info")
 
 # User management ################################################################ User management #
 api.add_resource(user.RetrieveUserInfo, "/user/info", endpoint="user_info")